cert_reqs accepts strings and numeric constants
by pushing the resolution of the argument down into ssl_wrap_socket the user is
able to alter the VerifiedHTTPSConnection.cert_reqs without in line with the
public API.
Setting it directly to a string wouldn't have worked earlier.

This also allows to use the constants defined in the ssl module as valid
parameter.