enable skip_host and skip_accept_encoding
Py2.4 added skip_host and skip_accept_encoding to httplib, but it's only
available when manually sending data over the wire with connect() and
putheader().

In httpconnection.py, httplib and http.client are wrapped with a small
fix to enable support for skip_host and skip_accept_encoding.

Also unit tests.