Added concept of a `RequestField`.
A `RequestField` is a data container the encapsulates what data is needed to
build post bodies of various forms. The `encode_multipart_formdata` function
was adjusted to always expect objects of type `RequestField`, and a backwards
compatible factory-style method was added to account for the old tuple-style
way of passing around the post body fields.

For inspiration of the change, see
https://github.com/shazow/urllib3/issues/215