urllib3 pyopenssl ssl_wrap_socket capability to override default _verify_callback
This is to allow customized server certificate validation.  Use cases include: a certificate that passes all of chain validation with the exception of the hostname (instances where the IP is specified during connection instead of the hostname) OR you wish to restrict connections to only servers providing specific certificates ie. a single allowable CN (common name).