Speed up tornado tests on Windows.
On Windows, socket.create_connection is delayed one second
if the host supports IPv6 and there is no listening IPv6 socket.
We fixed this with a updated version of bind_sockets which
uses the same port on IPv4 and IPv6 if automatic allocation is
requested.
See this upstream PR for details:
https://github.com/facebook/tornado/pull/977