Close a connection in the event of an SSLError
If an HTTPS connection is made with certificate verification turned on
and the verification of the certificate fails, the connection will be
returned to the connection pool with CERT_REQUIRED set and cannot be
reused. If we close and discard the connection we will avoid further
certificate verification errors.

The only potential problem I see with this approach is that if a user
wants to reconnect with verification but a different pem file, they'll
incur the cost of creating a new connection.