Seek body to original pos on retry. Fixes #459.
Fixes what I think is the typical case of using
a file-like object in HTTP PUT where urllib3
will retry the request.

Without the fix, when urllib3 retries a failed
PUT request, the body is not rewound and as
a result, if the request had a Content-Length,
the server would for bytes that are never sent
and time out on every retry until retry counter
runs out.