Fix check for chunked transfer-encoding
The Transfer-Encoding header can contain many values separated by
a comma, e.g.,

    Transfer-Encoding: chunked, gzip

If we want to check for chunked transfer encodings we're better off
not looking for equality since the above header value also indicates
a chunked body. Instead, we need to split on ", " to ensure we can
accurately determine if a response is chunked.