Issue #610 -- Tests fallback to IPv4 if IPv6 fails
On systems with IPv6 disabled all IPv6 unit tests will now be skipped,
instead of throwing an exception.

On some IPv6 enabled systems with badly configured DNS, where the IPv6
loopback address ::1 is not assigned to localhost, the unit tests will
fallback to using IPv4 for tests that bind to localhost instead of
throwing an socket.gaierror exception. IPv6 tests that bind directly to
::1 will continue to be run.