Suggest people `pip install cryptography`
When I tried to follow these instructions to use PyOpenSSL on Ubuntu 14.04, the `import urllib3.contrib.pyopenssl` line failed in Python as the `cryptography` was not already installed in my base Python.  Installing it separately (in the virtualenv in which I was doing this work) made it so these instructions worked.  I think it should be one of the requirements specified here; at worst, it will already be installed in the user's Python and be a no-op.

It was also necessary for me to `sudo apt-get install libffi-dev`, before some of the other `pip` requirements would install.  It may be worth noting that in a platform-agnostic way, or not.