Remove Python 3.2 support. Goodbye, old friend.
I know it's hard to understand, but sometimes painful things like this
happen. It's all part of the process of exploration and discovery. It's
all part of taking a chance and expanding man's horizons. The future
doesn't belong to the fainthearted; it belongs to the brave. Python 3.2
was pulling us into the future, and we'll continue to follow it.