Create connection pool keys by scheme
This commit modifies how the PoolManager stores and retrieves
``ConnectionPool`` objects for later use. It adds a new instance
variable to the class, ``pool_keys_by_scheme``, which allows a user to
define a list of fields from ``connection_pool_kw`` to consider (by
scheme) when storing and retrieving pools.