Remove no-longer accurate test.
This test no longer applies because we've unified the PyOpenSSL and
standard library logic. The standard library does not load the default
verify locations: that means, neither does PyOpenSSL anymore.

We can reconsider that decision (I quite like the logic used by
PyOpenSSL), but right now we're unifying the logic and so this test no
longer makes any sense.