``PoolManager`` now references the gobal ``key_fn_by_scheme``
Initially for #830, the ``key_fn_by_scheme`` was copied for each
PoolManager instance. This is inconcistent with the
``pool_classes_by_scheme`` instance variable, which is a reference to
the global dictionary. This commit makes the two consistent by making
the ``key_fn_by_scheme`` instance variable a reference to the global
dictionary.