updated SNIMissingWarning to address Jython
In my development environment, i need to disable SSL.  requests makes that easy by use using verify=False.  However, Jython doesn't allow SNI.  In order to disable it, I wrote the mentioned gist.  I'd be glad to make changes to it if anyone would find it helpful.  It was just a quick fix for me.