Backport Python 3.5 match_hostname function.
This backport actually comes from the repository maintained by Brandon
Rhodes: https://bitbucket.org/brandon/backports.ssl_match_hostname. The
reason for this is that the version in the Python 3.5 repository assumes
the presence of the ipaddress module: this version can function without
it. That makes this a good first step for handling IP address SANs, as
we can simply require that a backported ipaddress module be installed to
get the support required.