Add support for ChaCha20.
Currently we prefer AES GCM over ChaCha20, mostly because we suspect we
generally run on systems that have AES-NI extensions. We should keep an
eye on https://bugs.python.org/issue27766, as this issue adds a function
to the ssl module we can use to detect AES-NI's presence and optionally
use to get more info about what our priorities should be.