The ``connection_from_*`` methods now accept pool_kwargs
This introduces two changes. The first is that connection pools are
keyed off the entire request context rather than a sub-set of the
request context. The second is that it is possible to override
connection pool keyword arguments when calling ``connection_from_*`` by
providing a new keyword argument, ``pool_kwargs``. This is nice because
without it, users that want to change these arguments (when, say,
adjusting SSL configurations) need to make sure they update the keyword
arguments and get a new pool in a thread-safe manner.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>