Stop trying 'buffering' arg after TypeError
If `conn.getresponse(buffering=True)` throws TypeError once,
it's always going to throw it on all future calls.

The new HTTPConnectionPool._getresponse helper method added in
this patch instead replaces itself on the instance with the
underlying ConnectionCls method so all future calls on that
particular pool instance will use that instead of the pool
level helper method.

A runtime sanity check is included to detect the case where
a connection that isn't an instance of ConnectionCls is somehow
accessed through the pool.