Move retry on TimeoutError with method check to Retry class
The Retry class contains logic such as the method whitelist
and the functionality to actually increment the retries, so
it is more appropriate to have the check there. The tests
are also changed to accomodate this: the test is now
simplifed and it resides on the Retry test suite rather than
the ConnectionPool one. The test to check Retry's history
was changed such that it allows retries on POST: this is so
that it can pass without changing the previously tested history.