Remove all content length enforcement.
Sorry @nateprewitt, I know that you worked hard on this, but h11 handles
this for us more effectively than we can, so all of this code is
redundant with its own logic. It's good that you did it, though, becuase
it means our transition to being strict about content-length enforcement
is finally present!