Statically link OpenSSL and cryptography.
This is necessary to avoid a mess due to distutils trying to replicate
the compliation environment of the original Python. distutils loses
track of what it's trying to do and so compiles using the header files
from OpenSSL 1.1 but then links against OpenSSL 1.0, gets all in a
tizzy, and throws its hands up in the air and quits in a huff.

This avoids that by instead statically linking. It avoids getting the
Python linker flags passed through and instead just finds the archive
files directly. As an added benefit this is much more representative of
how the cryptography wheels are distributed, which means that we can
handily avoid testing in unrepresentative environments.

Good work team!