Stop buffering entire deflate-encoded responses
Previously, if we received a deflate-encoded response that included the
zlib header, we would buffer the entire response in case decompressing
failed and we wanted to try interpretting the response as a raw deflate
stream.

Now, we commit to a single decoder as soon as we've decoded any data. As
a result, we can drop the buffer at the same time.