Fix a tricky reentrancy bug in the Twisted backend
We used to assert that the event was gone after we ran the callback.
However, due to the way that Deferreds work, sometimes the event ended
up being un-registered and then re-registered again later by another
callback in the same chain.

This was discovered by trying to fetch http://httpbin.org/stream/10,
and finding that the Twisted backend crashed while the sync and Trio
backends worked. After this commit, all three backends now handle this
case correctly.

Thanks to @markrwilliams for figuring this out.