Expose full synchronous API to Python 3
Python 2 and Python 3 can now use the same synchronous API: in this API
there is no async color at all, which would be a syntax error in Python
3. This allows to remove the 'run secretly sync function as async' trick
which would have required to wrap the whole API for Python 3 sync. Since
we already have a solution for Python 2 sync, we use it for Python 3
too. As a result, the synchronous backend is never async colored, which
is why it is now in urllib3.backends instead of urllib3._async.backends.