Improve timeout handling in sync backend
urllib3 has quite an elaborate scheme to handle timeouts. None of that
is needed for trio, but it's needed for the sync version (and possibly
for other async frameworks, I'm not sure).

Anyway, we have a connect timeout (possibly set for each connection) and
a read timeout (possibly set for each request). The
connectionpool/request code takes care of sending the correct values to
each connection, what's left is passing this down to the backend and
making sure the backend handles that correctly.

Until now, both timeouts (connect and read) had issues because they were
set at backend creation time, which is too early. We now:

* add connect_timeout as a parameter to connect()
* add read_timeout as a parameter to send_request()