Handle EINTR from poll with zero timeout
While common sense says that it should be impossible, it seems that
poll() can indeed fail with EINTR on Linux even when the timeout is
zero. There was logic in _retry_on_intr which bypassed the restart
handling when the timeout was zero, which led to EINTRs leaking out.

Fixed by setting the deadline to 0 when the timeout is 0. This avoids an
extra system call to get monotonic(). It assumes that monotonic() is
always non-negative; that isn't documented but seems a reasonable
assumption given that the reference point tends to be something like
boot time.

Closes #1396.