Fix test_client_no_intermediate on macOS 10.13+
RFC 3280 and RFC 5280 specify that [serial numbers must not be longer
than 20 octets][0], but the serial number in client_no_intermediate.pem
was 38 bytes long. Using 20 bytes fixes the issue.

This test was passing on macOS 10.12. But since macOS 10.13, the serial
number are rejected when they're too long. macOS 10.12 was only warning
about the issue, see [SecCertificateParse][1]:

    /* RFC5280 4.1.2.2 limits serial number values to 20 octets.
       %%% At some point, this should be treated as a hard error.*/

But on macOS 10.13, [SecCertificateParse][2] now says:

    /* Note: RFC5280 4.1.2.2 limits serial number values to 20 octets.
       For now, we warn about larger values, but will still create the
       certificate with values up to 36 octets to avoid breaking some
       nonconforming certs with slightly longer serial numbers.
       We also explicitly allow serial numbers of 21 octets where the
       leading byte is 0x00 which is used to make a negative 20 octet
       value positive.
    */

It turns out that the [serial number in our certificate is 38 bytes
long][3], which is larger than the new 36 bytes limit. We want to use a
conforming serial number, which is why this new certificate uses 20
bytes.

The new certificate was generated using [trustme][4]. The long term plan
is to use trustme inside tests to generate all certificates, which will
allow to keep the generation of the certificate close to their usage,
and take advantage of future trustme improvements. However, trustme
does not support Python 2.6, so we'll need to drop Python 2.6 support in
urllib3 before switching to trustme.

test_client_intermediate also fails on macOS 10.13, but replacing the
certificate requires producing a certificate with an intermediate CA,
which is not possible yet.

Thanks to:

* @bartonjs for pointing out that the serial number length was the issue
* @yan12125 for showing log stream and the source of SecCertificateParse
* @njsmith for their help on using trustme

[0]: https://tools.ietf.org/html/rfc5280#section-4.1.2.2
[1]: https://opensource.apple.com/source/Security/Security-57740.60.18/OSX/sec/Security/SecCertificate.c.auto.html
[2]: https://opensource.apple.com/source/Security/Security-58286.41.2/OSX/sec/Security/SecCertificate.c.auto.html
[3]: https://stackoverflow.com/a/51194617
[4]: https://github.com/python-trio/trustme