Implement max_retry_wait_length for Retry objects
Retry object gets an extra parameter at initalization for max_retry_wait_length
and when it calls get_retry_after with a response that contains Retry-After header
it checks if the value of that greater than max_retry_wait_length. If it
is, then a ExceedingWaitTime exception will be raised.