Add CodeFund Sponsorship to README
CodeFund provides ethical sponsorships to open source maintainers. This PR will place the "Sponsored by" image at the top of the README. The sponsoring companies are not paying per click nor impression. They are paying the maintainer(s) on a per-month basis to be the primary sponsor of this project.

For an example, see https://github.com/gitcoinco/code_fund_ads/blob/master/README.md