Tell pytest to show native tracebacks
The default pytest tracebacks are really verbose: for each function in
the stack they include the whole function until the exception, and the
representation of its parameters.

On the other hand, native tracebacks are way more concise, showing only
two lines for each function and exception chains.

The benefits of switching to native tracebacks:

 * The useful information fits in one screen
 * Python trains us to read those tracebacks from day one
 * We avoid the verbosity of showing the source code: we can already
   see it in GitHub and our editors
 * While we lose the information about parameters, I've never found this
   information to be useful

I've been using this configuration for a few months now and never looked
back!