Document parameters to HTTPSConnection
This adds a small note in the docs that some of the params for
HTTPSConnection pass through to ssl_wrap_socket and leaves it at
that.

However, more directly useful for those trying to handle SNI by hand,
as part of the advanced-usage narrative docs this adds a new section.
It describes how to use SNI and connect to a host by IP. There might
be other usages like this which are interesting, but choosing the
address for a connection is a good, concrete example.

resolves #1884