Integrate SSLTransport into urllib3.
For connections that will attempt to use an HTTPS proxy with an HTTPS
destination, we'll use the TLS in TLS support provided by SSL Transport.

HTTPS proxy and HTTP destinations will continue using a single TLS
session as expected.

We'll still support the use of forwarding for HTTPS destinations with
HTTPS proxies as long as the "use_forwarding_for_https" parameter is
provided.

Signed-off-by: Jorge Lopez Silva <jalopezsilva@gmail.com>