Fix test_connection_cleanup_on_read_timeout flakiness
When calling `urlopen()`, even with `preload_content=False`, we read the
status line and headers, which can result in a read timeout. It happened
multiple times in CI since the read timeout is set to `SHORT_TIMEOUT`.

The fix is to simply allow those timeouts in the `urlopen()` call by
moving it inside the `pytest.raises()` context manager (and dropping
`preload_content=False`). This won't affect the test results as they're
about the connection pool: as long as we are able to connect, we're
good.

However, unlike `read()`, `urlopen()` can be retried on read timeouts,
which leads to the timeout exception being wrapped in `MaxRetryError` so
I had to disable that wrapping using `Retry(total=False)`.