Use system TLS ciphers if using OpenSSL 1.1.1+ or SecureTransport
OpenSSL 1.1.1+ supports its own list of pre-configured ciphers which are basically the same as our own list
so now we can stop setting ciphers out of fear that insecure ciphers will be used. This also means that
system administrators can configure their own ciphers in OpenSSL and we'll respect that automatically.

This also applies to SecureTransport which by default uses a good list of ciphers.