Better locate `method_whitelist` warning trigger
By default, `warnings.warn` will point to the source location of the warning
itself. For this case that is not very useful as the issue is in the caller
(whoever instantiates the `Retry` object). The `stacklevel` parameters allows
configuring how many stackframes up from the warning will be pointed to.