Lean on SSLContext to verify hostnames when possible (#2178)
* Use exact same SNI logic in pyOpenSSL and SecureTransport

It won't make a difference in practice, and avoids useless differences.

* Move "no SNI if IP address" responsibilty to TLS backends

 * The ssl module already avoids SNI when the host is an IP address
 * pyOpenSSL now does that too
 * We were already using SNI on IP addresses with SecureTransport anyway

In other words, this does not change anything, but is cleaner (we no
longer test for SecureTransport in ssl.py) and will allow us to lean on
ssl.SSLContext to match hostnames.

* Stop exposing is_ipaddress as a public

We already have way too much public functions

* Lean on SSLContext to verify hostnames when possible

* Address review comments