Change EmptyPoolError message
In blocking mode, if you get a connection from a pool and then try to
get another connection, EmptyPoolError is raised. Right now, it somewhat
confusingly says that the pool has reached maximum size, but in reality,
something else is happening.

The point of blocking mode is to guarantee that at most maxsize
connections are open at any time. Therefore, if the pool is empty, under
normal operation that means that a connection is currently in use and we
should not create a new connection.

I have updated the EmptyPoolError message to make this more clear.