Fix remove_headers_on_redirect feature modifying passed-in headers in-place
In general I'd say it's expected in Python that when passing a dict to
to an argument like `headers` that it won't be modified in-place and
could be reused safely for further requests.

Previously the remove_headers_on_redirect feature didn't uphold this. Do
a copy to make sure that it does.

This commit additionally *avoids* an unneeded copy when the
`remove_headers_on_redirect` code isn't reached.