Explicitly chain or don't chain exceptions
Explicitly chain exceptions (or disable the chaining) instead of
creating implicit "During handling of the above exception, another
exception occurred" which look like a bug.

Since urllib3 is now Python 3 only, we can use the built-in syntax for
this.

Methodology: grepped for `raise` and went over each case, deciding
whether to chain or `from None`. When in doubt I preferred to chain.

Fixes #2313.