Fix TypeError caused by interface mismatch
The PoolManager class takes the connection_pool_kw argument which
defines global parameters used for the connection pools managed by that
manager. Some of them are passed to the conn_kw argument to instantiate
a connection.

The problem here is that the PoolManager class instantiates different
classes of connection pools according to the protocol used. This
results in the issue caused by the different arguments in connection
classes(HTTPConnection vs HTTPSConnection).

This change introduces an additional kwargs to these connection classes
so that these classes can accept arguments which are not implemented.

Resolves: #2534