Fix HTTPSConnection/HTTPConnection reuse unapplied socket timeout bug
When making an HTTP(s) request to a host for which a connection has
already been established, this connection is reused however the timeout
provided is only applied to the HTTPSConnection/HTTPConnection object
but not it's underlying socket (the original requests timeout remains).
Therefore, if the initial request has a short timeout and a subsequent
request is initiated that happens to take more time than the initial,
the following error will be raised in accordance with the original
requests timeout:

     urllib3.exceptions.ProtocolError: ('Connection aborted.', TimeoutError('The write operation timed out'))

This commit fixes this be applying the timeout to the underlying socket.

Note this bug applied to HTTPConnections and HTTPSConnections using the
stdlib SSLContext backend but not HTTPSConnections using the PyOpenSSLContext
backend.